# swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Project Final",
    "version" : "v1",
    "title" : "ProjectFinal",
    "contact" : {
      "name" : "tesis",
      "email" : "leandro.cocchi@davinci.com"
    }
  },
  "host" : "*.*",
  "basePath" : "/",
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/issue" : {
      "get" : {
        "operationId" : "getAll",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Issue"
              }
            }
          }
        }
      },
      "post" : {
        "operationId" : "create",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Issue"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Issue"
            }
          }
        }
      }
    },
...
    "/issue/sprint/issues/{sprintId}" : {
      "get" : {
        "operationId" : "getIssuesBySprintId",
        "parameters" : [ {
          "name" : "sprintId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Issue"
              }
            }
          }
        }
      }
    },
    "/issue/sprint/{issueId}" : {
      "post" : {
        "operationId" : "setIssueInActiveSprint",
        "parameters" : [ {
          "name" : "issueId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Issue"
            }
          }
        }
      }
    },
    "/issue/sprint/{sprint}" : {
      "get" : {
        "operationId" : "getAllBySprint",
        "parameters" : [ {
          "name" : "sprint",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Issue"
              }
            }
          }
        }
      }
    },
    "/issue/{id}" : {
      "get" : {
        "operationId" : "getById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Issue"
            }
          }
        }
      },
      "put" : {
        "operationId" : "update",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Issue"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Issue"
            }
          }
        }
      },
      "delete" : {
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/label" : {
      "get" : {
        "operationId" : "getAll",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Label"
              }
            }
          }
        }
      },
      "post" : {
        "operationId" : "create",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Label"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Label"
            }
          }
        }
      }
    },
    "/label/{id}" : {
      "get" : {
        "operationId" : "getById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Label"
            }
          }
        }
      },
      "put" : {
        "operationId" : "update",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Label"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Label"
            }
          }
        }
      },
      "delete" : {
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/phase" : {
      "get" : {
        "operationId" : "getAll",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Phase"
              }
            }
          }
        }
      },
      "post" : {
        "operationId" : "create",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Phase"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Phase"
            }
          }
        }
      }
    },
    "/phase/{id}" : {
      "get" : {
        "operationId" : "getById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Phase"
            }
          }
        }
      },
      "put" : {
        "operationId" : "update",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Phase"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Phase"
            }
          }
        }
      },
      "delete" : {
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/rol/{id}" : {
      "get" : {
        "operationId" : "getById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Rol"
            }
          }
        }
      },
      "put" : {
        "operationId" : "update",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Rol"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Rol"
            }
          }
        }
      },
      "delete" : {
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/sprint/{id}" : {
      "get" : {
        "operationId" : "getById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Sprint"
            }
          }
        }
      },
      "put" : {
        "operationId" : "update",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Sprint"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Sprint"
            }
          }
        }
      },
      "delete" : {
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "Issue" : {
      "type" : "object",
      "required" : [ "backlog", "created", "enabled", "title", "updated" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "title" : {
          "type" : "string",
          "minLength" : 1
        },
        "avatar" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "priority" : {
          "type" : "string"
        },
        "label" : {
          "type" : "string"
        },
        "phase" : {
          "type" : "string"
        },
        "state" : {
          "type" : "string"
        },
        "version" : {
          "type" : "string"
        },
        "assignee" : {
          "type" : "string"
        },
        "reporter" : {
          "type" : "string"
        },
        "watcher" : {
          "type" : "integer",
          "format" : "int32"
        },
        "created" : {
          "type" : "string",
          "format" : "date-time"
        },
        "updated" : {
          "type" : "string",
          "format" : "date-time"
        },
        "resolved" : {
          "type" : "string",
          "format" : "date-time"
        },
        "plannedStart" : {
          "type" : "string",
          "format" : "date-time"
        },
        "plannedEnd" : {
          "type" : "string",
          "format" : "date-time"
        },
        "estimated" : {
          "type" : "integer",
          "format" : "int32"
        },
        "remaining" : {
          "type" : "integer",
          "format" : "int32"
        },
        "sprint" : {
          "type" : "integer",
          "format" : "int32"
        },
        "backlog" : {
          "type" : "boolean"
        },
        "enabled" : {
          "type" : "boolean"
        }
      }
    },
    "Label" : {
      "type" : "object",
      "required" : [ "created", "name" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "name" : {
          "type" : "string",
          "minLength" : 1
        },
        "description" : {
          "type" : "string"
        },
        "created" : {
          "type" : "string",
          "format" : "date-time"
        }
      }
    },
    "Phase" : {
      "type" : "object",
      "required" : [ "created", "name" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "name" : {
          "type" : "string",
          "minLength" : 1
        },
        "description" : {
          "type" : "string"
        },
        "created" : {
          "type" : "string",
          "format" : "date-time"
        }
      }
    },
...
    "Rol" : {
      "type" : "object",
      "required" : [ "created", "name" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "name" : {
          "type" : "string",
          "minLength" : 1
        },
        "description" : {
          "type" : "string"
        },
        "created" : {
          "type" : "string",
          "format" : "date-time"
        }
      }
    },
    "Sprint" : {
      "type" : "object",
      "required" : [ "created", "dateFrom", "dateTo", "enabled", "name" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "name" : {
          "type" : "string",
          "minLength" : 1
        },
        "description" : {
          "type" : "string"
        },
        "dateFrom" : {
          "type" : "string",
          "format" : "date-time"
        },
        "dateTo" : {
          "type" : "string",
          "format" : "date-time"
        },
        "created" : {
          "type" : "string",
          "format" : "date-time"
        },
        "enabled" : {
          "type" : "boolean"
        }
      }
    },
...
  }
}